Refactor octree collision primitives
